+++ dialect.go
@@ -94,6 +94,9 @@ func (dia DefaultDialect) CreateTable(model *Model, foreign map[string]ForeignKe
 		if sqlField.AutoIncrement() {
 			s += " AUTO_INCREMENT"
 		}
+		if _default := sqlField.Default(); _default != "" {
+			s += " DEFAULT " + _default
+		}
 		for k, v := range sqlField.Attrs() {
 			if v == "" {
 				s += " " + k