+++ dialect_sqlite.go
@@ -50,6 +50,9 @@ func (dia Sqlite3Dialect) CreateTable(model *Model, foreign map[string]ForeignKe
 		if sqlField.AutoIncrement() {
 			s += " AUTOINCREMENT"
 		}
+		if _default := sqlField.Default(); _default != "" {
+			s += " DEFAULT " + _default
+		}
 		for k, v := range sqlField.Attrs() {
 			if v == "" {
 				s += " " + k